Fixing _plustic materials
Materials named as Material_plustic do not have any ingot for input on the forging anvil due to the ore dictionary not finding anything in the ingotMaterial_Plustic oredict (because it ddoesn't exist).  Naming it exist and adding an igot to it still does not make the anvil find the ingot.  To work around this issue, I recreated the offending materials via Pewter and gave them the appropriate name so the anvil can find the input ingot.